requirements.txt
pyodbc==4.0.39
pandas==2.1.4
python-dotenv==1.0.0
pyyaml==6.0.1 